requirements.txt
asgiref==3.5.0
boto3==1.21.29
botocore==1.24.29
Django==3.0.14
django-cors-headers==3.11.0
django-filter==21.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jmespath==1.0.0
Pillow==9.0.1
psycopg2-binary==2.8.6
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==0.14.1
whitenoise==3.3.1
